Get service account projects from Machines, not Cluster.
When creating a service account for the machine-controller, rather than
requiring a Project be defined in the Cluster object ProviderConfig,
instead look at the ProviderConfigs of all of the Machine objects, and
add permissions for them all for the service account.

Also, change from using []Machine to []*Machine most places to play
nicely with other code.