Start using Cluster API type
This installs the cluster CRD and puts a cluster object in the API
server. The master IP is now stored in the cluter status and passed to
the things that need it, so a lot of cleanup around that. This also
fixes the client creation in the machine controller to use the global
scheme so that the parameter codec for used for list options is aware
of our API group.