Use version prefix matching for Debian packages.
When Kubernetes creates Debian packages for a release, they add a suffix
to the package version, so "1.8.0" actually becomes "1.8.0-00". This
allows for patching the package and creating new Debian releases based
off of the same Kubernetes release (e.g. "1.8.0-01", "1.8.0-02", etc) in
the event that there was something wrong with the packaging, but not
Kubernetes itself.

However, not every Kubernetes version has an -00 package (if they are
particularly troublesome, they get removed from the repo), and if there
are multiple such versions available, we should try to use the newest
one.

This change uses prefix matching to find the best package version for
each component (given the Kubernetes SemVer), instead of always
hardcoding the "-00" suffix.